T-Shirt Shop with SQLite Database
Recently, I finished a course with my teacher Sérgio Luiz Banin, and I wanted to create a simple project using the concepts I learned.

This is a basic Python console application that simulates a t-shirt store. It allows the user to choose between different brands and sizes, and it records the purchase data in a local SQLite database.

This project was built to reinforce concepts such as:
- Working with SQLite in Python
- Using the `sqlite3` and `time` libraries
- Handling user input
- Structuring code in a readable way

Feel free to clone, explore or improve it!